airdrops work

+++ client/src/lib.ts
@@ -18,20 +18,20 @@ import * as crypto from 'crypto';
 import { personAccSize } from './instancedata';
 
 
-export const createUserKey = async (filePath: string): Promise<string> => {
-    const newPair = Keypair.generate();
-    console.log(`New key stored at ${filePath}, public key ${newPair.publicKey}`);
-    fs.writeFile(filePath, `[${newPair.secretKey.toString()}]`, { encoding: 'utf8' });
-    return filePath;
+export const storeRandomKeypair = async (filePath: string, overwrite = false): Promise<Keypair> => {
+    const fileExists = await fs.exists(filePath)
+    if (!fileExists) {
+        const newPair = Keypair.generate();
+        await fs.writeFile(filePath, `[${newPair.secretKey.toString()}]`, { encoding: 'utf8' });
+    }
+    const secretKeyString = await fs.readFile(filePath, { encoding: 'utf8' });
+    const secretKey = Uint8Array.from(JSON.parse(secretKeyString));
+    const ret = Keypair.fromSecretKey(secretKey);
+    console.log(`Key stored at ${filePath}, public key ${ret.publicKey}`);
+    return ret;
+
 };
 
-export const getWalletKeyPair = async (userPath: string): Promise<Keypair> => {
-    const secretKeyString = await fs.readFile(userPath, { encoding: 'utf8' });
-    const secretKey = Uint8Array.from(JSON.parse(secretKeyString));
-    const keyPair = Keypair.fromSecretKey(secretKey);
-    console.log(`Wallet found at ${keyPair.publicKey.toString()}`)
-    return keyPair;
-};
 
 export const createPersonAccount = async (connection: Connection, payer: Keypair, programId: PublicKey) => {
     const FIXED_ACC_SEED = 'A unique seed that would normally be securely generated';
@@ -88,3 +88,24 @@ export const getProgramKeypair = async (connection: Connection, programPath: str
     return programKeyPair;
 };
 
+/**
+ * 
+ * @param connection 
+ * @param keypair Account to airdrop funds to
+ * @param multipleOfMin 
+ * @returns Account balance after airdrop
+ */
+export const airdropMin = async (connection: Connection, keypair: Keypair, multipleOfMin = 2): Promise<number> => {
+    // Minimum size per Solana docs https://docs.solana.com/developing/programming-model/accounts
+    const MIN_ACC_BYTES = 128;
+    const MIN_EXEMPT_LAMPORTS = await connection.getMinimumBalanceForRentExemption(
+        MIN_ACC_BYTES,
+    );
+    const AIRDROP_AMOUNT = multipleOfMin * MIN_EXEMPT_LAMPORTS;
+    console.log('amount', AIRDROP_AMOUNT);
+    const airdropTx = await connection.requestAirdrop(keypair.publicKey, AIRDROP_AMOUNT);
+    await connection.confirmTransaction(airdropTx);
+
+    const balance = await connection.getBalance(keypair.publicKey);
+    return balance;
+}
